Adapt remaining files in src.
