+++ src/globals.d.ts
@@ -1,0 +1,2 @@
+// https://github.com/storybookjs/storybook/issues/2883
+declare module "*.md";
