+++ src/onepirate/ForgotPassword.tsx
@@ -3,7 +3,6 @@ import withRoot from "./modules/withRoot";
 import React from "react";
 import { Field, Form, FormSpy } from "react-final-form";
 import { makeStyles } from "@material-ui/core/styles";
-import Link from "@material-ui/core/Link";
 import Typography from "./modules/components/Typography";
 import AppFooter from "./modules/views/AppFooter";
 import AppAppBar from "./modules/views/AppAppBar";
@@ -12,7 +11,6 @@ import { email, required } from "./modules/form/validation";
 import RFTextField from "./modules/form/RFTextField";
 import FormButton from "./modules/form/FormButton";
 import FormFeedback from "./modules/form/FormFeedback";
-import { Link as RouterLink } from "react-router-dom";
 
 const useStyles = makeStyles((theme) => ({
   form: {
@@ -27,17 +25,17 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
-function SignIn() {
+function ForgotPassword() {
   const classes = useStyles();
   const [sent, setSent] = React.useState(false);
 
-  const validate = (values) => {
+  const validate = (values: { [index: string]: string }) => {
     const errors = required(["email", "password"], values);
 
     if (!errors.email) {
-      const emailError = email(values.email, values);
+      const emailError = email(values.email);
       if (emailError) {
-        errors.email = email(values.email, values);
+        errors.email = email(values.email);
       }
     }
 
@@ -54,18 +52,11 @@ function SignIn() {
       <AppForm>
         <React.Fragment>
           <Typography variant="h3" gutterBottom marked="center" align="center">
-            Sign In
+            Forgot your password?
           </Typography>
           <Typography variant="body2" align="center">
-            {"Not a member yet? "}
-            <Link
-              component={RouterLink}
-              to="/sign-up/"
-              align="center"
-              underline="always"
-            >
-              Sign Up here
-            </Link>
+            {"Enter your email address below and we'll " +
+              "send you a link to reset your password."}
           </Typography>
         </React.Fragment>
         <Form
@@ -73,11 +64,11 @@ function SignIn() {
           subscription={{ submitting: true }}
           validate={validate}
         >
-          {({ handleSubmit2, submitting }) => (
+          {({ handleSubmit: handleSubmit2, submitting }) => (
             <form onSubmit={handleSubmit2} className={classes.form} noValidate>
               <Field
-                autoComplete="email"
                 autoFocus
+                autoComplete="email"
                 component={RFTextField}
                 disabled={submitting || sent}
                 fullWidth
@@ -86,18 +77,6 @@ function SignIn() {
                 name="email"
                 required
                 size="large"
-              />
-              <Field
-                fullWidth
-                size="large"
-                component={RFTextField}
-                disabled={submitting || sent}
-                required
-                name="password"
-                autoComplete="current-password"
-                label="Password"
-                type="password"
-                margin="normal"
               />
               <FormSpy subscription={{ submitError: true }}>
                 {({ submitError }) =>
@@ -115,24 +94,15 @@ function SignIn() {
                 color="secondary"
                 fullWidth
               >
-                {submitting || sent ? "In progress…" : "Sign In"}
+                {submitting || sent ? "In progress…" : "Send reset link"}
               </FormButton>
             </form>
           )}
         </Form>
-        <Typography align="center">
-          <Link
-            underline="always"
-            component={RouterLink}
-            to="/forgot-password/"
-          >
-            Forgot password?
-          </Link>
-        </Typography>
       </AppForm>
       <AppFooter />
     </React.Fragment>
   );
 }
 
-export default withRoot(SignIn);
+export default withRoot(ForgotPassword);
