+++ src/onepirate/modules/form/FormButton.tsx
@@ -1,13 +1,16 @@
 import React from "react";
 import Button from "../components/Button";
 import defer from "./defer";
+import { ButtonProps } from "@material-ui/core";
 
 interface FormButtonProps {
   disabled?: boolean;
   mounted?: boolean;
 }
 
-function FormButton(props: FormButtonProps) {
+function FormButton<C extends React.ElementType>(
+  props: FormButtonProps & ButtonProps<C, { component?: C }>
+) {
   const { disabled, mounted, ...others } = props;
   return (
     <Button
